+++ workspaces/src/network/server.rs
@@ -1,22 +1,49 @@
 use std::fs::File;
+use std::net::{Ipv4Addr, SocketAddrV4};
+use std::path::PathBuf;
 
 use crate::error::{ErrorKind, SandboxErrorCode};
 use crate::result::Result;
+use crate::types::SecretKey;
 
 use async_process::Child;
 use fs2::FileExt;
-use portpicker::pick_unused_port;
+use near_account_id::AccountId;
+use reqwest::Url;
 use tempfile::TempDir;
 use tracing::info;
 
 use near_sandbox_utils as sandbox;
+use tokio::net::TcpListener;
+
+// Must be an IP address as `neard` expects socket address for network address.
+const DEFAULT_RPC_HOST: &str = "127.0.0.1";
+
+fn rpc_socket(port: u16) -> String {
+    format!("{DEFAULT_RPC_HOST}:{}", port)
+}
+
+/// Request an unused port from the OS.
+pub async fn pick_unused_port() -> Result<u16> {
+    // Port 0 means the OS gives us an unused port
+    // Important to use localhost as using 0.0.0.0 leads to users getting brief firewall popups to
+    // allow inbound connections on MacOS.
+    let addr = SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0);
+    let listener = TcpListener::bind(addr)
+        .await
+        .map_err(|err| ErrorKind::Io.full("failed to bind to random port", err))?;
+    let port = listener
+        .local_addr()
+        .map_err(|err| ErrorKind::Io.full("failed to get local address for random port", err))?
+        .port();
+    Ok(port)
+}
 
 /// Acquire an unused port and lock it for the duration until the sandbox server has
 /// been started.
-fn acquire_unused_port() -> Result<(u16, File)> {
+async fn acquire_unused_port() -> Result<(u16, File)> {
     loop {
-        let port = pick_unused_port()
-            .ok_or_else(|| SandboxErrorCode::InitFailure.message("no ports free"))?;
+        let port = pick_unused_port().await?;
         let lockpath = std::env::temp_dir().join(format!("near-sandbox-port{}.lock", port));
         let lockfile = File::create(lockpath).map_err(|err| {
             ErrorKind::Io.full(format!("failed to create lockfile for port {}", port), err)
@@ -39,38 +66,86 @@ async fn init_home_dir() -> Result<TempDir> {
     Ok(home_dir)
 }
 
-pub struct SandboxServer {
-    pub(crate) rpc_port: u16,
-    pub(crate) net_port: u16,
-    pub(crate) home_dir: TempDir,
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum ValidatorKey {
+    HomeDir(PathBuf),
+    Known(AccountId, SecretKey),
+}
 
+pub struct SandboxServer {
+    pub(crate) validator_key: ValidatorKey,
+
+    rpc_addr: Url,
+    net_port: Option<u16>,
     rpc_port_lock: Option<File>,
     net_port_lock: Option<File>,
     process: Option<Child>,
 }
 
 impl SandboxServer {
+    /// Connect a sandbox server that's already been running, provided we know the rpc_addr
+    /// and home_dir pointing to the sandbox process.
+    pub(crate) async fn connect(rpc_addr: String, validator_key: ValidatorKey) -> Result<Self> {
+        let rpc_addr = Url::parse(&rpc_addr).map_err(|e| {
+            SandboxErrorCode::InitFailure.full(format!("Invalid rpc_url={rpc_addr}"), e)
+        })?;
+        Ok(Self {
+            validator_key,
+            rpc_addr,
+            net_port: None,
+            rpc_port_lock: None,
+            net_port_lock: None,
+            process: None,
+        })
+    }
+
+    /// Run a new SandboxServer, spawning the sandbox node in the process.
     pub(crate) async fn run_new() -> Result<Self> {
         // Supress logs for the sandbox binary by default:
         supress_sandbox_logs_if_required();
 
-        // Try running the server with the follow provided rpc_ports and net_ports
-        let (rpc_port, rpc_port_lock) = acquire_unused_port()?;
-        let (net_port, net_port_lock) = acquire_unused_port()?;
-        let home_dir = init_home_dir().await?;
-
+        let home_dir = init_home_dir().await?.into_path();
         // Configure `$home_dir/config.json` to our liking. Sandbox requires extra settings
         // for the best user experience, and being able to offer patching large state payloads.
         crate::network::config::set_sandbox_configs(&home_dir)?;
-        let child = sandbox::run(&home_dir, rpc_port, net_port)
+
+        // Try running the server with the follow provided rpc_ports and net_ports
+        let (rpc_port, rpc_port_lock) = acquire_unused_port().await?;
+        let (net_port, net_port_lock) = acquire_unused_port().await?;
+        // It's important that the address doesn't have a scheme, since the sandbox expects
+        // a valid socket address.
+        let rpc_addr = rpc_socket(rpc_port);
+        let net_addr = rpc_socket(net_port);
+
+        info!(target: "workspaces", "Starting up sandbox at localhost:{}", rpc_port);
+
+        let options = &[
+            "--home",
+            home_dir
+                .as_os_str()
+                .to_str()
+                .expect("home_dir is valid utf8"),
+            "run",
+            "--rpc-addr",
+            &rpc_addr,
+            "--network-addr",
+            &net_addr,
+        ];
+
+        let child = sandbox::run_with_options(options)
             .map_err(|e| SandboxErrorCode::RunFailure.custom(e))?;
 
         info!(target: "workspaces", "Started up sandbox at localhost:{} with pid={:?}", rpc_port, child.id());
 
+        let rpc_addr: Url = format!("http://{rpc_addr}")
+            .parse()
+            .expect("static scheme and host name with variable u16 port numbers form valid urls");
+
         Ok(Self {
-            rpc_port,
-            net_port,
-            home_dir,
+            validator_key: ValidatorKey::HomeDir(home_dir),
+            rpc_addr,
+            net_port: Some(net_port),
             rpc_port_lock: Some(rpc_port_lock),
             net_port_lock: Some(net_port_lock),
             process: Some(child),
@@ -83,7 +158,10 @@ impl SandboxServer {
         if let Some(rpc_port_lock) = self.rpc_port_lock.take() {
             rpc_port_lock.unlock().map_err(|e| {
                 ErrorKind::Io.full(
-                    format!("failed to unlock lockfile for rpc_port={}", self.rpc_port),
+                    format!(
+                        "failed to unlock lockfile for rpc_port={:?}",
+                        self.rpc_port()
+                    ),
                     e,
                 )
             })?;
@@ -91,7 +169,7 @@ impl SandboxServer {
         if let Some(net_port_lock) = self.net_port_lock.take() {
             net_port_lock.unlock().map_err(|e| {
                 ErrorKind::Io.full(
-                    format!("failed to unlock lockfile for net_port={}", self.net_port),
+                    format!("failed to unlock lockfile for net_port={:?}", self.net_port),
                     e,
                 )
             })?;
@@ -100,8 +178,16 @@ impl SandboxServer {
         Ok(())
     }
 
+    pub fn rpc_port(&self) -> Option<u16> {
+        self.rpc_addr.port()
+    }
+
+    pub fn net_port(&self) -> Option<u16> {
+        self.net_port
+    }
+
     pub fn rpc_addr(&self) -> String {
-        format!("http://localhost:{}", self.rpc_port)
+        self.rpc_addr.to_string()
     }
 }
 
@@ -111,12 +197,13 @@ impl Drop for SandboxServer {
             return;
         }
 
+        let rpc_port = self.rpc_port();
         let child = self.process.as_mut().unwrap();
 
         info!(
             target: "workspaces",
-            "Cleaning up sandbox: port={}, pid={}",
-            self.rpc_port,
+            "Cleaning up sandbox: port={:?}, pid={}",
+            rpc_port,
             child.id()
         );
 
@@ -124,9 +211,6 @@ impl Drop for SandboxServer {
             .kill()
             .map_err(|e| format!("Could not cleanup sandbox due to: {:?}", e))
             .unwrap();
-
-        // Unlock the ports just in case they have not been preemptively done.
-        self.unlock_lockfiles().unwrap();
     }
 }